feat(html-reporter): add labels to test file and test case - separate labels into their own component - fixint tests
