+++ packages/html-reporter/src/testCaseView.tsx
@@ -23,7 +23,7 @@ import { ProjectLink } from './links';
 import { statusIcon } from './statusIcon';
 import './testCaseView.css';
 import { TestResultView } from './testResultView';
-import { LabelsView, matchTags } from './labelsView';
+import { hashStringToInt, matchTags } from './labelUtils';
 
 export const TestCaseView: React.FC<{
   projectNames: string[],
@@ -51,7 +51,7 @@ export const TestCaseView: React.FC<{
     {test && <div className='test-case-location'>{test.location.file}:{test.location.line}</div>}
     {test && (!!test.projectName || labels) && <div className='test-case-project-labels-row'>
       {!!test.projectName && <ProjectLink projectNames={projectNames} projectName={test.projectName}></ProjectLink>}
-      {labels && <LabelsView labels={labels} />}
+      {labels && <LabelsLinkView labels={labels} />}
     </div>}
     {annotations.size > 0 && <AutoChip header='Annotations'>
       {[...annotations].map(annotation => <TestCaseAnnotationView type={annotation[0]} descriptions={annotation[1]} />)}
@@ -93,3 +93,19 @@ function retryLabel(index: number) {
     return 'Run';
   return `Retry #${index}`;
 }
+
+const LabelsLinkView: React.FC<React.PropsWithChildren<{
+  labels: string[],
+}>> = ({ labels }) => {
+  return labels.length > 0 ? (
+    <>
+      {labels.map(tag => (
+        <a style={{ textDecoration: 'none', color: 'var(--color-fg-default)' }} href={`#?q=@${tag}`} >
+          <span style={{ margin: '6px 0 0 6px', cursor: 'pointer' }} className={'label label-color-' + (hashStringToInt(tag))}>
+            {tag}
+          </span>
+        </a>
+      ))}
+    </>
+  ) : null;
+};
